src/shipit_risk_assessment/requirements_frozen.txt
aioamqp==0.10.0
aiohttp==2.1.0
appdirs==1.4.3
async-timeout==1.2.1
certifi==2017.4.17
chardet==3.0.4
click==6.7
flake8==3.3.0
flake8-coding==1.3.0
flake8-quotes==0.11.0
idna==2.5
Logbook==1.0.0
mccabe==0.6.1
mohawk==0.3.4
mozilla-cli-common==1.0.0
mozilla-mocoda==0.1.0
multidict==2.1.6
packaging==16.8
py==1.4.34
pycodestyle==2.3.1
pyflakes==1.5.0
pyparsing==2.2.0
pytest==3.1.1
python-hglib==2.4
raven==6.1.0
requests==2.17.3
six==1.10.0
slugid==1.0.7
structlog==17.2.0
taskcluster==1.3.3
urllib3==1.21.1
whatthepatch==0.0.5
yarl==0.10.2
